Spring Boot

+++ src/main/java/com/ust/pos/user/UserController.java
@@ -4,6 +4,8 @@ import com.ust.pos.dto.UserDto;
 import com.ust.pos.role.service.RoleService;
 import com.ust.pos.user.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
@@ -11,23 +13,25 @@ import org.springframework.web.bind.annotation.*;
 @Controller
 @RequestMapping("/user")
 public class UserController {
-    @Autowired
-    private RoleService roleService;
 
     @Autowired
     private UserService userService;
 
+    @Autowired
+    private RoleService roleService;
+
     @GetMapping("/list")
     public String home(Model model) {
-        model.addAttribute("user", userService.findAll());
+        model.addAttribute("users", userService.findAll());
+        model.addAttribute("roles", roleService.findAll());
         return "user/list";
     }
 
     @GetMapping("/get")
     public String update(Model model, @RequestParam String username) {
-        UserDto response = userService.findByUserName(username);
+        UserDto userDto = userService.findByUserName(username);
+        model.addAttribute("userDto", userDto);
         model.addAttribute("roles", roleService.findAll());
-        model.addAttribute("user", response);
         return "user/user";
     }
 
@@ -36,7 +40,7 @@ public class UserController {
         UserDto response = userService.update(userDto);
         if (!response.isSuccess()) {
             model.addAttribute("message", response.getMessage());
-            model.addAttribute("user", userDto);
+            model.addAttribute("userDto", userDto);
             return "user/user";
         }
         return "redirect:/user/list";
@@ -44,7 +48,15 @@ public class UserController {
 
     @GetMapping("/delete")
     public String delete(Model model, @RequestParam String username) {
-        userService.delete(username);
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if(authentication != null){
+            String loggedInUser = authentication.getName();
+            userService.delete(username);
+            if (loggedInUser.equals(username)) {
+                SecurityContextHolder.clearContext();
+                return "redirect:/login";
+            }
+        }
         return "redirect:/user/list";
     }
 }
